+++ CHANGELOG.md
@@ -44,4 +44,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `bnb.choice(...parsers)`
+- Method `parser.skip`
+- Method `parser.next`
+- Function `bnb.choice(...parsers)`